requirements.txt
pygame
PyYaml
rlbot
tqdm
# wandb
plotly
kaleido
argparse
einops
rlgym_compat
pandas